refactor(api): split provider/mod.rs into registry and factory modules
